sample_code/neuro-demo/server/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
cloud-volume
pyspng-seunglab
python-multipart
psutil==5.9.8